# LITHOVEX/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LITHOVEX AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: AI chat operations
  - name: chats
    description: Chat history operations
  - name: search
    description: Web search operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chat/completions:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/search:
    post:
      operationId: webSearch
      tags: [search]
      summary: Perform a web search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List all saved chats
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryList"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewChatPayload"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistory"

  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a chat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateChat
      tags: [chats]
      summary: Update a chat (title, messages)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditChatPayload"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      required:
        - role
        - content

    ChatRequest:
      type: object
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        webSearchEnabled:
          type: boolean
        projectContext:
          type: ["string", "null"]
      required:
        - model
        - messages

    ChatResponse:
      type: object
      properties:
        content:
          type: string
        model:
          type: string
        tokensUsed:
          type: ["integer", "null"]
      required:
        - content
        - model

    SearchRequest:
      type: object
      properties:
        query:
          type: string
        maxResults:
          type: integer
      required:
        - query

    SearchResult:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        href:
          type: string
      required:
        - title
        - body
        - href

    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        query:
          type: string
      required:
        - results
        - query

    ChatHistory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - messages
        - createdAt
        - updatedAt

    ChatHistoryList:
      type: object
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistory"
      required:
        - chats

    NewChatPayload:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    EditChatPayload:
      type: object
      properties:
        title:
          type: ["string", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error